requirements.txt
aiofiles==0.8.0
cffi==1.15.1
httptools==0.4.0
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.4
multidict==6.0.2
py-algorand-sdk==1.15.0
pycparser==2.21
pycryptodomex==3.15.0
PyJWT==2.4.0
PyNaCl==1.5.0
sanic==22.6.0
sanic-routing==22.3.0
ujson==5.4.0
uvloop==0.16.0
websockets==10.3
